+++ packages/combobox/src/Combobox/Combobox.tsx
@@ -131,6 +131,7 @@ export function Combobox<M extends boolean>({
   usePortal = true,
   portalClassName,
   portalContainer,
+  portalRef,
   scrollContainer,
   popoverZIndex,
   ...rest
@@ -1160,6 +1161,7 @@ export function Combobox<M extends boolean>({
           usePortal,
           portalClassName,
           portalContainer,
+          portalRef,
           scrollContainer,
         }
       : { usePortal }),
@@ -1427,6 +1429,12 @@ Combobox.propTypes = {
   usePortal: PropTypes.bool,
   scrollContainer: PropTypes.elementType,
   portalContainer: PropTypes.elementType,
+  portalRef: PropTypes.shape({
+    current:
+      typeof window !== 'undefined'
+        ? PropTypes.instanceOf(Element)
+        : PropTypes.any,
+  }),
   portalClassName: PropTypes.string,
 };