+++ packages/modal/src/index.ts
@@ -1,3 +1,4 @@
+export * from './constants';
 export { Footer } from './Footer';
 export {
   CloseIconColor,