+++ packages/select/src/ListMenu/ListMenu.tsx
@@ -36,6 +36,7 @@ const ListMenu = React.forwardRef<HTMLUListElement, ListMenuProps>(
       dropdownWidthBasis,
       usePortal = true,
       portalContainer,
+      portalRef,
       scrollContainer,
       portalClassName,
       popoverZIndex,
@@ -77,7 +78,13 @@ const ListMenu = React.forwardRef<HTMLUListElement, ListMenuProps>(
       onExit,
       onExited,
       ...(usePortal
-        ? { usePortal, portalClassName, portalContainer, scrollContainer }
+        ? {
+            usePortal,
+            portalClassName,
+            portalContainer,
+            portalRef,
+            scrollContainer,
+          }
         : { usePortal }),
     };