+++ packages/tabs/src/context/index.ts
@@ -1,0 +1,9 @@
+export {
+  TabDescendantsContext,
+  useTabDescendantsContext,
+} from './TabDescendantsContext';
+export {
+  TabPanelDescendantsContext,
+  useTabPanelDescendantsContext,
+} from './TabPanelDescendantsContext';
+export { TabsContext, TabsContextProps, useTabsContext } from './TabsContext';