+++ packages/tabs/src/utils/getEnabledIndices/getEnabledIndices.ts
@@ -1,0 +1,5 @@
+export function getEnabledIndices(tabTitleElements: Array<HTMLElement>) {
+  return tabTitleElements
+    .filter(tabTitleEl => !tabTitleEl.hasAttribute('disabled'))
+    .map(tabTitleEl => tabTitleElements.indexOf(tabTitleEl));
+}