+++ tools/codemods/src/utils/jsx/insertJSXComment/tests/transform.spec.ts
@@ -1,0 +1,6 @@
+import { transformTest } from '../../../tests/transformTest';
+
+transformTest(__dirname, {
+  fixture: 'insert-jsx-comment',
+  transform: 'insert-jsx-comment',
+});